+++ test/tap/async-context.tap.ts
@@ -9,7 +9,7 @@ test("asynchronously propagating state with local-context-domains", function (t)
   t.plan(2);
 
   const namespace = cls.createNamespace("namespace");
-  t.ok(process.namespaces.namespace, "namespace has been created");
+  t.ok(cls.getNamespaces().namespace, "namespace has been created");
 
   namespace.run(function () {
     namespace.set("test", 1337);